Rename Option to ArgOption and outsource With* functions
